Move and rename random map with flags
